Refactored code to deal with messages being sent on command.
Messages get added to a string of messages to send —> then get sent as
that single string in the communication thread function (now called
“TalkToClient()”). This potentially introduces a small delay between
when the message is told to be sent and actually sent.

May have to refactor or take timestamps into account later. Not totally
clear on what this timing is supposed to be like for the actual
hardware & synchronization.